restructured ivshmem subdirectories
