+++ configure.ac
@@ -104,9 +104,9 @@ AM_CONDITIONAL([OPENIB_ENABLED],
 # -- IVSHmem --
 
 AC_ARG_ENABLE([ivshmem],[AS_HELP_STRING([--disable-ivshmem],
-	      [disable InterVMsharedMemory OFED with RC support])])
+	      [disable inter-VM shared memory support])])
 AS_IF([test "x$enable_ivshmem" != xno],
-      [AX_CHECK_LIBRARY([IVSHMEM], [infiniband/verbs.h], [ibverbs],
+      [AX_CHECK_LIBRARY([IVSHMEM], [pthread.h], [pthread],
       [have_ivshmem=yes], [have_ivshmem=no])])
 AS_IF([test "x$have_ivshmem" = xno -a "x$enable_ivshmem" = xyes],
       [AC_MSG_ERROR([ivshmem support requested but not found])])
@@ -217,7 +217,6 @@ AC_CONFIG_FILES(lib/pscom4ofed/Makefile)
 AC_CONFIG_FILES(lib/pscom4ucp/Makefile)
 AC_CONFIG_FILES(lib/pscom4openib/Makefile)
 AC_CONFIG_FILES(lib/pscom4ivshmem/Makefile)  #Added for ivshmem plugin
-AC_CONFIG_FILES(lib/psmalloc4ivshmem/Makefile) #Added for ivshmem direct mode
 AC_CONFIG_FILES(lib/pscom4psm/Makefile)
 AC_CONFIG_FILES(lib/psmalloc/Makefile)
 AC_CONFIG_FILES(lib/psport4/Makefile)
