+++ lib/pscom/Makefile.am
@@ -26,7 +26,7 @@ libpscom_la_SOURCES = pscom.c pscom_con.c pscom_debug.c			\
 	pscom_plugin.c pscom_queues.c pscom_req.c pscom_shm.c		\
 	pscom_sock.c pscom_str_util.c pscom_tcp.c pscom_ufd.c pslib.c	\
 	psshmalloc.c pscom_precon.c pscom_async.c pscom_suspend.c       \
-	pscom_migrate.c psivshmem_malloc.c ../pscom4ivshmem/psivshmem.c
+	pscom_migrate.c
 
 noinst_HEADERS = getid.c list.h pscom_con.h pscom_debug.h pscom_env.h	\
 	pscom_gm.h pscom_group.h pscom_io.h pscom_p4s.h			\
