+++ lib/pscom/Makefile.in
@@ -153,7 +153,6 @@ am__v_lt_ = $(am__v_lt_@AM_DEFAULT_V@)
 am__v_lt_0 = --silent
 am__v_lt_1 = 
 libpscom_la_DEPENDENCIES =
-am__dirstamp = $(am__leading_dot)dirstamp
 am_libpscom_la_OBJECTS = libpscom_la-pscom.lo libpscom_la-pscom_con.lo \
 	libpscom_la-pscom_debug.lo libpscom_la-pscom_dprint.lo \
 	libpscom_la-pscom_env.lo libpscom_la-pscom_group.lo \
@@ -166,8 +165,7 @@ am_libpscom_la_OBJECTS = libpscom_la-pscom.lo libpscom_la-pscom_con.lo \
 	libpscom_la-pscom_ufd.lo libpscom_la-pslib.lo \
 	libpscom_la-psshmalloc.lo libpscom_la-pscom_precon.lo \
 	libpscom_la-pscom_async.lo libpscom_la-pscom_suspend.lo \
-	libpscom_la-pscom_migrate.lo libpscom_la-psivshmem_malloc.lo \
-	../pscom4ivshmem/libpscom_la-psivshmem.lo
+	libpscom_la-pscom_migrate.lo
 libpscom_la_OBJECTS = $(am_libpscom_la_OBJECTS)
 libpscom_la_LINK = $(LIBTOOL) $(AM_V_lt) --tag=CC $(AM_LIBTOOLFLAGS) \
 	$(LIBTOOLFLAGS) --mode=link $(CCLD) $(libpscom_la_CFLAGS) \
@@ -438,7 +436,7 @@ libpscom_la_SOURCES = pscom.c pscom_con.c pscom_debug.c			\
 	pscom_plugin.c pscom_queues.c pscom_req.c pscom_shm.c		\
 	pscom_sock.c pscom_str_util.c pscom_tcp.c pscom_ufd.c pslib.c	\
 	psshmalloc.c pscom_precon.c pscom_async.c pscom_suspend.c       \
-	pscom_migrate.c psivshmem_malloc.c ../pscom4ivshmem/psivshmem.c
+	pscom_migrate.c
 
 noinst_HEADERS = getid.c list.h pscom_con.h pscom_debug.h pscom_env.h	\
 	pscom_gm.h pscom_group.h pscom_io.h pscom_p4s.h			\
@@ -539,15 +537,6 @@ clean-noinstLTLIBRARIES:
 
 libperf.la: $(libperf_la_OBJECTS) $(libperf_la_DEPENDENCIES) $(EXTRA_libperf_la_DEPENDENCIES) 
 	$(AM_V_CCLD)$(LINK)  $(libperf_la_OBJECTS) $(libperf_la_LIBADD) $(LIBS)
-../pscom4ivshmem/$(am__dirstamp):
-	@$(MKDIR_P) ../pscom4ivshmem
-	@: > ../pscom4ivshmem/$(am__dirstamp)
-../pscom4ivshmem/$(DEPDIR)/$(am__dirstamp):
-	@$(MKDIR_P) ../pscom4ivshmem/$(DEPDIR)
-	@: > ../pscom4ivshmem/$(DEPDIR)/$(am__dirstamp)
-../pscom4ivshmem/libpscom_la-psivshmem.lo:  \
-	../pscom4ivshmem/$(am__dirstamp) \
-	../pscom4ivshmem/$(DEPDIR)/$(am__dirstamp)
 
 libpscom.la: $(libpscom_la_OBJECTS) $(libpscom_la_DEPENDENCIES) $(EXTRA_libpscom_la_DEPENDENCIES) 
 	$(AM_V_CCLD)$(libpscom_la_LINK) -rpath $(libdir) $(libpscom_la_OBJECTS) $(libpscom_la_LIBADD) $(LIBS)
@@ -563,13 +552,10 @@ libpslib.la: $(libpslib_la_OBJECTS) $(libpslib_la_DEPENDENCIES) $(EXTRA_libpslib
 
 mostlyclean-compile:
 	-rm -f *.$(OBJEXT)
-	-rm -f ../pscom4ivshmem/*.$(OBJEXT)
-	-rm -f ../pscom4ivshmem/*.lo
 
 distclean-compile:
 	-rm -f *.tab.c
 
-@AMDEP_TRUE@@am__include@ @am__quote@../pscom4ivshmem/$(DEPDIR)/libpscom_la-psivshmem.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libpscom_la-pscom.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libpscom_la-pscom_async.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libpscom_la-pscom_con.Plo@am__quote@
@@ -593,7 +579,6 @@ distclean-compile:
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libpscom_la-pscom_suspend.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libpscom_la-pscom_tcp.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libpscom_la-pscom_ufd.Plo@am__quote@
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libpscom_la-psivshmem_malloc.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libpscom_la-pslib.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/libpscom_la-psshmalloc.Plo@am__quote@
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/perf.Plo@am__quote@
@@ -800,26 +785,11 @@ libpscom_la-pscom_migrate.lo: pscom_migrate.c
 @AMDEP_TRUE@@am__fastdepCC_FALSE@	DEPDIR=$(DEPDIR) $(CCDEPMODE) $(depcomp) @AMDEPBACKSLASH@
 @am__fastdepCC_FALSE@	$(AM_V_CC@am__nodep@)$(LIBTOOL) $(AM_V_lt) --tag=CC $(AM_LIBTOOLFLAGS) $(LIBTOOLFLAGS) --mode=compile $(CC) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) $(libpscom_la_CFLAGS) $(CFLAGS) -c -o libpscom_la-pscom_migrate.lo `test -f 'pscom_migrate.c' || echo '$(srcdir)/'`pscom_migrate.c
 
-libpscom_la-psivshmem_malloc.lo: psivshmem_malloc.c
-@am__fastdepCC_TRUE@	$(AM_V_CC)$(LIBTOOL) $(AM_V_lt) --tag=CC $(AM_LIBTOOLFLAGS) $(LIBTOOLFLAGS) --mode=compile $(CC) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) $(libpscom_la_CFLAGS) $(CFLAGS) -MT libpscom_la-psivshmem_malloc.lo -MD -MP -MF $(DEPDIR)/libpscom_la-psivshmem_malloc.Tpo -c -o libpscom_la-psivshmem_malloc.lo `test -f 'psivshmem_malloc.c' || echo '$(srcdir)/'`psivshmem_malloc.c
-@am__fastdepCC_TRUE@	$(AM_V_at)$(am__mv) $(DEPDIR)/libpscom_la-psivshmem_malloc.Tpo $(DEPDIR)/libpscom_la-psivshmem_malloc.Plo
-@AMDEP_TRUE@@am__fastdepCC_FALSE@	$(AM_V_CC)source='psivshmem_malloc.c' object='libpscom_la-psivshmem_malloc.lo' libtool=yes @AMDEPBACKSLASH@
-@AMDEP_TRUE@@am__fastdepCC_FALSE@	DEPDIR=$(DEPDIR) $(CCDEPMODE) $(depcomp) @AMDEPBACKSLASH@
-@am__fastdepCC_FALSE@	$(AM_V_CC@am__nodep@)$(LIBTOOL) $(AM_V_lt) --tag=CC $(AM_LIBTOOLFLAGS) $(LIBTOOLFLAGS) --mode=compile $(CC) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) $(libpscom_la_CFLAGS) $(CFLAGS) -c -o libpscom_la-psivshmem_malloc.lo `test -f 'psivshmem_malloc.c' || echo '$(srcdir)/'`psivshmem_malloc.c
-
-../pscom4ivshmem/libpscom_la-psivshmem.lo: ../pscom4ivshmem/psivshmem.c
-@am__fastdepCC_TRUE@	$(AM_V_CC)$(LIBTOOL) $(AM_V_lt) --tag=CC $(AM_LIBTOOLFLAGS) $(LIBTOOLFLAGS) --mode=compile $(CC) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) $(libpscom_la_CFLAGS) $(CFLAGS) -MT ../pscom4ivshmem/libpscom_la-psivshmem.lo -MD -MP -MF ../pscom4ivshmem/$(DEPDIR)/libpscom_la-psivshmem.Tpo -c -o ../pscom4ivshmem/libpscom_la-psivshmem.lo `test -f '../pscom4ivshmem/psivshmem.c' || echo '$(srcdir)/'`../pscom4ivshmem/psivshmem.c
-@am__fastdepCC_TRUE@	$(AM_V_at)$(am__mv) ../pscom4ivshmem/$(DEPDIR)/libpscom_la-psivshmem.Tpo ../pscom4ivshmem/$(DEPDIR)/libpscom_la-psivshmem.Plo
-@AMDEP_TRUE@@am__fastdepCC_FALSE@	$(AM_V_CC)source='../pscom4ivshmem/psivshmem.c' object='../pscom4ivshmem/libpscom_la-psivshmem.lo' libtool=yes @AMDEPBACKSLASH@
-@AMDEP_TRUE@@am__fastdepCC_FALSE@	DEPDIR=$(DEPDIR) $(CCDEPMODE) $(depcomp) @AMDEPBACKSLASH@
-@am__fastdepCC_FALSE@	$(AM_V_CC@am__nodep@)$(LIBTOOL) $(AM_V_lt) --tag=CC $(AM_LIBTOOLFLAGS) $(LIBTOOLFLAGS) --mode=compile $(CC) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) $(CPPFLAGS) $(libpscom_la_CFLAGS) $(CFLAGS) -c -o ../pscom4ivshmem/libpscom_la-psivshmem.lo `test -f '../pscom4ivshmem/psivshmem.c' || echo '$(srcdir)/'`../pscom4ivshmem/psivshmem.c
-
 mostlyclean-libtool:
 	-rm -f *.lo
 
 clean-libtool:
 	-rm -rf .libs _libs
-	-rm -rf ../pscom4ivshmem/.libs ../pscom4ivshmem/_libs
 
 ID: $(am__tagged_files)
 	$(am__define_uniq_tagged_files); mkid -fID $$unique
@@ -936,8 +906,6 @@ clean-generic:
 distclean-generic:
 	-test -z "$(CONFIG_CLEAN_FILES)" || rm -f $(CONFIG_CLEAN_FILES)
 	-test . = "$(srcdir)" || test -z "$(CONFIG_CLEAN_VPATH_FILES)" || rm -f $(CONFIG_CLEAN_VPATH_FILES)
-	-rm -f ../pscom4ivshmem/$(DEPDIR)/$(am__dirstamp)
-	-rm -f ../pscom4ivshmem/$(am__dirstamp)
 
 maintainer-clean-generic:
 	@echo "This command is intended for maintainers to use"
@@ -948,7 +916,7 @@ clean-am: clean-generic clean-libLTLIBRARIES clean-libtool \
 	clean-noinstLTLIBRARIES mostlyclean-am
 
 distclean: distclean-am
-	-rm -rf ../pscom4ivshmem/$(DEPDIR) ./$(DEPDIR)
+	-rm -rf ./$(DEPDIR)
 	-rm -f Makefile
 distclean-am: clean-am distclean-compile distclean-generic \
 	distclean-tags
@@ -995,7 +963,7 @@ install-ps-am:
 installcheck-am:
 
 maintainer-clean: maintainer-clean-am
-	-rm -rf ../pscom4ivshmem/$(DEPDIR) ./$(DEPDIR)
+	-rm -rf ./$(DEPDIR)
 	-rm -f Makefile
 maintainer-clean-am: distclean-am maintainer-clean-generic
