+++ lib/pscom4ivshmem/Makefile.am
@@ -15,12 +15,16 @@ include $(top_srcdir)/common.am
 AM_LDFLAGS += -avoid-version
 AM_CPPFLAGS += -DLIBDIR=\"$(libdir)\" -I$(top_srcdir)/lib/pscom
 
-lib_LTLIBRARIES = libpscom4ivshmem.la
+lib_LTLIBRARIES = libpscom4ivshmem.la libpsmalloc4ivshmem.la
 
-libpscom4ivshmem_la_SOURCES = pscom_ivshmem.c pscom_ivshmem.h psivshmem.c psivshmem.h metadata.h ##psivshmem_malloc.c psivshmem_malloc.h
+libpscom4ivshmem_la_SOURCES = pscom_ivshmem.c pscom_ivshmem.h psivshmem.c psivshmem.h metadata.h ./direct_mode/psivshmem_malloc.c ./direct_mode/psivshmem_malloc.h
 libpscom4ivshmem_la_CPPFLAGS = $(AM_CPPFLAGS) $(IVSHMEM_CPPFLAGS)
 libpscom4ivshmem_la_LDFLAGS = $(AM_LDFLAGS) $(IVSHMEM_LDFLAGS)
 libpscom4ivshmem_la_LIBADD =  -lpthread $(top_builddir)/lib/pscom/libpscom.la
+libpsmalloc4ivshmem_la_SOURCES = ./direct_mode/psmallocivshmem.c ./direct_mode/psivshmem_malloc.h ./direct_mode/psivshmem_malloc.c psivshmem.c psivshmem.h
+libpsmalloc4ivshmem_la_CPPFLAGS = $(AM_CPPFLAGS) $(PSM_CPPFLAGS)
+libpsmalloc4ivshmem_la_LDFLAGS = $(AM_LDFLAGS) $(PSM_LDFLAGS)
+libpsmalloc4ivshmem_la_LIBADD = $(top_builddir)/lib/pscom/libpscom.la
 
 
 install-exec-hook:
